btree/tests: Move iterator tests to a dedicated file
